requirements.txt
scikit-learn==1.5.1
pandas==2.2.2
jupyterlab==4.2.3
fastapi==0.111.0
uvicorn==0.30.1
joblib==1.4.2
streamlit==1.36.0
xgboost==2.1.0
hyperopt==0.2.7
lightgbm==4.4.0
lime==0.2.0.1
wandb==0.17.4
requests==2.32.3
python-dotenv==1.0.1
plotly==5.19.0